FIX: Read Actions API

+++ src/main/java/org/antonakospanos/iot/atlas/dao/model/ConditionConjunctive.java
@@ -37,9 +37,9 @@ public class ConditionConjunctive implements Serializable {
 
 	private String value;
 
-	private Integer minValue;
+	private Double minValue;
 
-	private Integer maxValue;
+	private Double maxValue;
 
 
 	public Long getId() {
@@ -82,19 +82,19 @@ public class ConditionConjunctive implements Serializable {
 		this.value = value;
 	}
 
-	public Integer getMinValue() {
+	public Double getMinValue() {
 		return minValue;
 	}
 
-	public void setMinValue(Integer minValue) {
+	public void setMinValue(Double minValue) {
 		this.minValue = minValue;
 	}
 
-	public Integer getMaxValue() {
+	public Double getMaxValue() {
 		return maxValue;
 	}
 
-	public void setMaxValue(Integer maxValue) {
+	public void setMaxValue(Double maxValue) {
 		this.maxValue = maxValue;
 	}
 }
